fastapi-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
authlib==1.2.1
httpx==0.25.1
itsdangerous==2.1.2
pydantic==2.4.2
email-validator==2.1.0.post1